Crate filetype

This crate provides a basic extension to std::fs::File: it defines a method which returns the file's type (on *nix systems).

use std::fs;
use std::fs::File;
use filetype::{FileType, UnixFileType};

let f = File::open("foo").unwrap();
let ftype = f.file_type().unwrap();

match ftype {
    FileType::Regular => {},
    FileType::Directory => {
    },
    FileType::Symlink => {
    },
    FileType::NamedPipe => {
    },
    FileType::BlockDevice => {
    },
    FileType::CharacterDevice => {
    },
}

Enums

Error
An error which occurred during file type determination. Either an error returned by nix::sys::stat::fstat, or the file type is uknown. If unknown, the file mask is returned; the file mask is the st_mode returned by stat bitwise-ANDed with libc::consts::os::posix88::S_IFMT. See the man pages for fstat for more information.

FileType
Every standard Unix file type (except for Sockets, since this is not provided by libc::consts::os::posix88)

Traits

UnixFileType
Can return a Unix file type.

Type Definitions

FileTypeResult